move slider from form-validation module to slider module
