+++ js/main.js
@@ -3,3 +3,4 @@ import './data.js';
 import './page.js';
 import './ad-form.js';
 import  './map.js';
+import './slider.js';
